+++ src/components/Registration/CompanySelectBottomSheet.tsx
@@ -1,0 +1,46 @@
+const CompanySelectBottomSheet = () => {
+  return (
+    <div className="modal-dimmed">
+      <div className="modal">
+        <div className="flex-center">
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+        </div>
+        <div className="flex-center">
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린 카드</span>
+          </div>
+          <div className="modal-item-container">
+            <div className="modal-item-dot"></div>
+            <span className="modal-item-name">클린카드</span>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default CompanySelectBottomSheet;
