+++ src/pages/CardListPage/index.tsx
@@ -7,7 +7,7 @@ const CardListPage = () => {
   return (
     <div className="flex-column-center">
       <CardListTitle>보유 카드</CardListTitle>
-      <Card />
+      {/**카드 */}
       <AddCardButton onClick={() => navigate('/registration')} />
     </div>
   );
@@ -18,33 +18,6 @@ const CardListTitle = ({ children }: { children: ReactNode }) => {
     <div className="flex-center">
       <h2 className="page-title mb-10">{children}</h2>
     </div>
-  );
-};
-
-const Card = () => {
-  return (
-    <>
-      <div className="card-box">
-        <div className="small-card">
-          <div className="card-top">
-            <span className="card-text">클린카드</span>
-          </div>
-          <div className="card-middle">
-            <div className="small-card__chip"></div>
-          </div>
-          <div className="card-bottom">
-            <div className="card-bottom__number">
-              <span className="card-text">1111 - 2222 - oooo - oooo</span>
-            </div>
-            <div className="card-bottom__info">
-              <span className="card-text">YUJO</span>
-              <span className="card-text">12 / 23</span>
-            </div>
-          </div>
-        </div>
-      </div>
-      <span className="card-nickname">법인카드</span>
-    </>
   );
 };
